// src/app/models/product.model.ts
// export class Product {
//   constructor(
//     public productId: number,
//     public productName: string,
//     public price: number,
//     public imagePath: string
//   ) {}
// }

export class Product {
  productId: number;
  productName: string;
  price: number;
  imagePath: string;
  quatity?: number;
}